+++ praxis.hub/src/net/neilcsmith/praxis/hub/CreationControl.java
@@ -46,7 +46,7 @@ class CreationControl extends BasicControl {
     private DefaultHub hub;
     private ControlInfo info;
     private Queue<Call> callQueue;
-    private long matchID;
+    private int matchID;
     private ControlAddress address;
     private ComponentFactory factory;
